+++ services/app-api/handlers/templates/fetch.test.ts
@@ -1,4 +1,4 @@
-import { getTemplate } from "./get";
+import { fetchTemplate } from "./fetch";
 import { APIGatewayProxyEvent } from "aws-lambda";
 import { proxyEvent } from "../../utils/testing/proxyEvent";
 import { StatusCodes } from "../../utils/types/types";
@@ -24,7 +24,7 @@ const testEvent: APIGatewayProxyEvent = {
   pathParameters: { templateName: "test" },
 };
 
-describe("Test getTemplate API method", () => {
+describe("Test fetchTemplate API method", () => {
   beforeAll(() => {
     process.env["TEMPLATE_BUCKET"] = "fakeTestBucket";
   });
@@ -34,7 +34,7 @@ describe("Test getTemplate API method", () => {
       ...testEvent,
       pathParameters: { templateName: "MCPAR" },
     };
-    const res = await getTemplate(mcparEvent, null);
+    const res = await fetchTemplate(mcparEvent, null);
 
     expect(res.statusCode).toBe(StatusCodes.SUCCESS);
     expect(res.body).toContain("s3://fakeurl.bucket.here");
@@ -45,7 +45,7 @@ describe("Test getTemplate API method", () => {
       ...testEvent,
       pathParameters: { templateName: "MLR" },
     };
-    const res = await getTemplate(mlrEvent, null);
+    const res = await fetchTemplate(mlrEvent, null);
 
     expect(res.statusCode).toBe(StatusCodes.SUCCESS);
     expect(res.body).toContain("s3://fakeurl.bucket.here");
@@ -56,7 +56,7 @@ describe("Test getTemplate API method", () => {
       ...testEvent,
       pathParameters: { templateName: "NAAAR" },
     };
-    const res = await getTemplate(naaarEvent, null);
+    const res = await fetchTemplate(naaarEvent, null);
 
     expect(res.statusCode).toBe(StatusCodes.SUCCESS);
     expect(res.body).toContain("s3://fakeurl.bucket.here");
@@ -67,7 +67,7 @@ describe("Test getTemplate API method", () => {
       ...testEvent,
       pathParameters: {},
     };
-    const res = await getTemplate(noKeyEvent, null);
+    const res = await fetchTemplate(noKeyEvent, null);
 
     expect(res.statusCode).toBe(500);
     expect(res.body).toContain(error.NO_TEMPLATE_NAME);
@@ -78,7 +78,7 @@ describe("Test getTemplate API method", () => {
       ...testEvent,
       pathParameters: { templateName: "wrongName" },
     };
-    const res = await getTemplate(noKeyEvent, null);
+    const res = await fetchTemplate(noKeyEvent, null);
 
     expect(res.statusCode).toBe(500);
     expect(res.body).toContain(error.INVALID_TEMPLATE_NAME);
